requirements.txt
aws-cdk-lib==2.63.0
aws-cdk.aws-apigatewayv2-alpha==2.63.0a0
aws-cdk.aws-apigatewayv2-authorizers-alpha==2.63.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.63.0a0
aws-cdk.aws-lambda-python-alpha==2.63.0a0
aws-lambda-powertools==2.7.1
behave==1.2.6
boto3-stubs==1.26.50
boto3==1.20.32
botocore==1.23.32
constructs>=10.0.0,<11.0.0
moto==4.1.0
mypy-boto3-dynamodb==1.26.24
mypy-boto3-lambda==1.26.49
PyJWT==2.6.0
pytest==7.2.0
pytest-cov==4.0.0
requests==2.28.2
